requirements.txt
xt
# requirements_driver.txt
# Versions compatibles testées pour Python 3.11
numpy==1.25.2
pandas==2.1.4
opencv-python==4.8.1.78
matplotlib==3.8.0
pillow==11.0.0
scipy==1.11.4
torch==2.1.0
torchvision==0.16.0
albumentations==1.4.2
scikit-learn==1.3.2
tqdm==4.66.1